--- data/src/main/java/com/example/compmoviedb/data/models/movieactors/Cast.kt
@@ -1,16 +1,0 @@
-package com.example.compmoviedb.data.models.movieactors
-
-data class Cast(
-    val adult: Boolean?,
-    val cast_id: Int?,
-    val character: String?,
-    val credit_id: String?,
-    val gender: Int?,
-    val id: Int?,
-    val known_for_department: String?,
-    val name: String?,
-    val order: Int?,
-    val original_name: String?,
-    val popularity: Double?,
-    val profile_path: String?
-)
