--- data/src/main/java/com/example/compmoviedb/data/models/movieactors/ListActorsMovieEntity.kt
@@ -1,7 +1,0 @@
-package com.example.compmoviedb.data.models.movieactors
-
-data class ListActorsMovieEntity(
-    val cast: List<Cast>?,
-    val crew: List<Crew>?,
-    val id: Int?
-)
