--- data/src/main/java/com/example/compmoviedb/data/models/moviedetails/BelongsToCollection.kt
@@ -1,8 +1,0 @@
-package com.example.compmoviedb.data.models.moviedetails
-
-data class BelongsToCollection(
-    val backdrop_path: String?,
-    val id: Int?,
-    val name: String?,
-    val poster_path: String?
-)
