--- data/src/main/java/com/example/compmoviedb/data/models/moviedetails/Genre.kt
@@ -1,6 +1,0 @@
-package com.example.compmoviedb.data.models.moviedetails
-
-data class Genre(
-    val id: Int?,
-    val name: String?
-)
