--- data/src/main/java/com/example/compmoviedb/data/models/moviedetails/ProductionCompany.kt
@@ -1,8 +1,0 @@
-package com.example.compmoviedb.data.models.moviedetails
-
-data class ProductionCompany(
-    val id: Int?,
-    val logo_path: String?,
-    val name: String?,
-    val origin_country: String?
-)
