--- data/src/main/java/com/example/compmoviedb/data/models/moviedetails/ProductionCountry.kt
@@ -1,6 +1,0 @@
-package com.example.compmoviedb.data.models.moviedetails
-
-data class ProductionCountry(
-    val iso_3166_1: String?,
-    val name: String?
-)
