--- data/src/main/java/com/example/compmoviedb/data/models/moviedetails/SpokenLanguage.kt
@@ -1,6 +1,0 @@
-package com.example.compmoviedb.data.models.moviedetails
-
-data class SpokenLanguage(
-    val iso_639_1: String?,
-    val name: String?
-)
