--- data/src/main/java/com/example/compmoviedb/data/models/moviespopular/ListMoviesPopularEntity.kt
@@ -1,8 +1,0 @@
-package com.example.compmoviedb.data.models.moviespopular
-
-data class ListMoviesPopularEntity(
-    val page: Int,
-    val results: List<Result>,
-    val total_pages: Int,
-    val total_results: Int
-)
