--- data/src/main/java/com/example/compmoviedb/data/models/movievideo/MovieVideoEntity.kt
@@ -1,6 +1,0 @@
-package com.example.compmoviedb.data.models.movievideo
-
-data class MovieVideoEntity(
-    val id: Int?,
-    val results: List<Result>?
-)
