--- data/src/main/java/com/example/compmoviedb/data/models/movievideo/Result.kt
@@ -1,14 +1,0 @@
-package com.example.compmoviedb.data.models.movievideo
-
-data class Result(
-    val id: String?,
-    val iso_3166_1: String?,
-    val iso_639_1: String?,
-    val key: String?,
-    val name: String?,
-    val official: Boolean?,
-    val published_at: String?,
-    val site: String?,
-    val size: Int?,
-    val type: String?
-)
