+++ data/src/main/java/com/example/compmoviedb/data/models/remote/movieactors/Crew.kt
@@ -1,0 +1,15 @@
+package com.example.compmoviedb.data.models.remote.movieactors
+
+data class Crew(
+    val adult: Boolean?,
+    val credit_id: String?,
+    val department: String?,
+    val gender: Int?,
+    val id: Int?,
+    val job: String?,
+    val known_for_department: String?,
+    val name: String?,
+    val original_name: String?,
+    val popularity: Double?,
+    val profile_path: String?
+)
