+++ data/src/main/java/com/example/compmoviedb/data/models/remote/movieactors/ListActorsMovieEntity.kt
@@ -1,0 +1,7 @@
+package com.example.compmoviedb.data.models.remote.movieactors
+
+data class ListActorsMovieEntity(
+    val cast: List<Cast>?,
+    val crew: List<Crew>?,
+    val id: Int?
+)
