+++ data/src/main/java/com/example/compmoviedb/data/models/remote/moviedetails/BelongsToCollection.kt
@@ -1,0 +1,8 @@
+package com.example.compmoviedb.data.models.remote.moviedetails
+
+data class BelongsToCollection(
+    val backdrop_path: String?,
+    val id: Int?,
+    val name: String?,
+    val poster_path: String?
+)
