+++ data/src/main/java/com/example/compmoviedb/data/models/remote/moviedetails/ProductionCompany.kt
@@ -1,0 +1,8 @@
+package com.example.compmoviedb.data.models.remote.moviedetails
+
+data class ProductionCompany(
+    val id: Int?,
+    val logo_path: String?,
+    val name: String?,
+    val origin_country: String?
+)
