+++ data/src/main/java/com/example/compmoviedb/data/models/remote/moviedetails/ProductionCountry.kt
@@ -1,0 +1,6 @@
+package com.example.compmoviedb.data.models.remote.moviedetails
+
+data class ProductionCountry(
+    val iso_3166_1: String?,
+    val name: String?
+)
