+++ data/src/main/java/com/example/compmoviedb/data/models/remote/moviedetails/SpokenLanguage.kt
@@ -1,0 +1,6 @@
+package com.example.compmoviedb.data.models.remote.moviedetails
+
+data class SpokenLanguage(
+    val iso_639_1: String?,
+    val name: String?
+)
