+++ data/src/main/java/com/example/compmoviedb/data/models/remote/movievideo/MovieVideoEntity.kt
@@ -1,0 +1,6 @@
+package com.example.compmoviedb.data.models.remote.movievideo
+
+data class MovieVideoEntity(
+    val id: Int?,
+    val results: List<Result>?
+)
