+++ data/src/main/java/com/example/compmoviedb/data/remote/dataSource/themoviedb/MovieDBApiInterface.kt
@@ -1,9 +1,9 @@
 package com.example.compmoviedb.data.remote.dataSource.themoviedb
 
-import com.example.compmoviedb.data.models.movieactors.ListActorsMovieEntity
-import com.example.compmoviedb.data.models.moviedetails.MovieDetailsEntity
-import com.example.compmoviedb.data.models.moviespopular.ListMoviesPopularEntity
-import com.example.compmoviedb.data.models.movievideo.MovieVideoEntity
+import com.example.compmoviedb.data.models.remote.movieactors.ListActorsMovieEntity
+import com.example.compmoviedb.data.models.remote.moviedetails.MovieDetailsEntity
+import com.example.compmoviedb.data.models.remote.moviespopular.ListMoviesPopularEntity
+import com.example.compmoviedb.data.models.remote.movievideo.MovieVideoEntity
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
